compression: check the length of the ROHC buffer before writing in it
Not all compression/encoding functions check for the length of the ROHC buffer
before writing ROHC bytes in it. Add some more checks. Some cases without check
probably still exist, especially in the TCP profile.

+++ src/comp/schemes/cid.h
@@ -36,11 +36,11 @@
  * Prototypes of functions that may used by other ROHC modules
  */
 
-size_t ROHC_EXPORT code_cid_values(const rohc_cid_type_t cid_type,
-                                   const int cid,
-                                   unsigned char *const dest,
-                                   const size_t dest_size,
-                                   size_t *const first_position)
+int ROHC_EXPORT code_cid_values(const rohc_cid_type_t cid_type,
+                                const int cid,
+                                unsigned char *const dest,
+                                const size_t dest_size,
+                                size_t *const first_position)
 	__attribute__((warn_unused_result, nonnull(3, 5)));
